refactor(import): move font import logic out of widget layer
Extract import analysis, assignment, and shared import models into util.
Keep widget files focused on dialog rendering and event wiring.
